backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQLdb==2.0.0
Werkzeug==3.0.4
itsdangerous==2.2.0
click==8.1.7
Jinja2==3.1.4
PyPDF2==3.0.1
python-docx==1.1.2
